requirements.txt
matplotlib
pandas
google-generativeai
google-api-python-client